requirements.txt
flask
flask-cors
opencv-python-headless
numpy
pillow
pytesseract
defusedxml
lxml
mammoth
openpyxl
